animation_orchestrator/requirements.txt
google-genai
fastapi
uvicorn
python-multipart
pillow
rembg[cpu]
opencv-python-headless
numpy
requests
pydantic
google-auth
firebase-admin
torch
torchvision
coloredlogs
platformdirs
sympy
flatbuffers
packaging
stripe
